feat : add api

+++ BE/contacts/api.py
@@ -7,17 +7,14 @@ from .serializers import FriendADDSerializer, FriendListSerializer
 from rest_framework.permissions import AllowAny #일시적으로 인증 비활성화, 개발 중에만 사용
 
 #친구 추가 API / 수정사항 필요, 이미 추가 되어있는 친구라면 안되도록 해야함
-#파티룸 주소 필드 수정 / 디데이 계산하는 기능 
+#파티룸 주소 필드 수정
 class FriendAddAPI (APIView):
 
-    permission_classes = [AllowAny] #
+    #permission_classes = [AllowAny]
 
-    def post(self,request,owner_id):
+    def post(self,request):
 
-        try:
-            owner = CustomUser.objects.get(id=owner_id)
-        except CustomUser.DoesNotExist :
-            return Response({'error':'owner가 존재하지 않습니다'},status=status.HTTP_404_NOT_FOUND)
+        Owner = request.user
         
         serializer_username = FriendADDSerializer(data=request.data)
         serializer_username.is_valid(raise_exception=True)
@@ -29,10 +26,9 @@ class FriendAddAPI (APIView):
             return Response({'error':'해당하는 친구가 존재하지 않습니다'},status=status.HTTP_404_NOT_FOUND)
 
         ADD_friend = Contact(
-            owner = owner,
+            owner = Owner,
             username=SEARCH_friend.username,
             birthday=SEARCH_friend.birthday,
-            d_day = 10,
             #party_room_address = 
         )
         ADD_friend.save()
@@ -42,24 +38,27 @@ class FriendAddAPI (APIView):
 #친구 리스트 조회 API
 class FriendListAPI (APIView):
 
-    permission_classes = [AllowAny]#
-    
-    def get (self,request,owner_id):
-        owner = CustomUser.objects.get(id=owner_id)
-        queryset = Contact.objects.filter(owner=owner)
+    #permission_classes = [AllowAny]
+    def get (self,request):
+        Owner = request.user
+        queryset = Contact.objects.filter(owner=Owner)
         serializer = FriendListSerializer(queryset,many=True)
         return Response(serializer.data,status=status.HTTP_200_OK)
     
-#친구 삭제 API // 로그인 구현 후 다시 구현
+#친구 삭제 API
 class FriendDeleteAPI (APIView):
     
-    permission_classes = [AllowAny] #실제로는 로그인이 되어있어야함. 본인의 친구창에서 지워야하기 때문
+    #permission_classes = [AllowAny]
 
     def delete (self,request,friend_id):
-        
+        Owner = request.user
         try :
             friend = Contact.objects.get(id=friend_id)
         except Contact.DoesNotExist:
             return Response({'error':'해당하는 친구가 없습니다'},status=status.HTTP_404_NOT_FOUND)
-        friend.delete()
-        return Response({'success':'친구 삭제 완료'},status=status.HTTP_204_NO_CONTENT)   
+        
+        if friend.owner == Owner:
+            friend.delete()
+            return Response({'success':'친구 삭제 완료'},status=status.HTTP_204_NO_CONTENT)  
+        else:
+            return  Response({'message':'잘못된 접근입니다'},status=status.HTTP_403_FORBIDDEN)
